Refactor into multiple files
